WIP - move extractor to dedicated crate
