+++ crates/indexify_extractor/Cargo.toml
@@ -1,0 +1,22 @@
+[package]
+name = "indexify_extractor"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+[lib]
+crate-type = ["rlib"]
+
+[dependencies]
+anyhow = { workspace = true }
+askama = { workspace = true }
+async-trait = { workspace = true }
+bollard = { workspace = true }
+pyo3 = { workspace = true }
+indexify_internal_api = { workspace = true }
+indexify_api = { workspace = true }
+mime = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+tokio-stream = { workspace = true }
+tracing = { workspace = true }
