+++ crates/indexify_extractor/src/python_path.rs
