+++ crates/indexify_extractor/src/scaffold.rs
