+++ src/extractor.rs
@@ -1,0 +1,2 @@
+/// Moved to crates/indexify_extractor
+pub use indexify_extractor::*;
