+++ src/main.rs
@@ -17,7 +17,6 @@ pub mod state;
 pub mod task_allocator;
 
 mod api;
-mod api_utils;
 mod blob_storage;
 mod caching;
 mod cmd;
